remove bambam::model::state module (redudant, exists in bambam_core)
