+++ rust/bambam/src/model/constraint/multimodal/engine.rs
@@ -1,6 +1,7 @@
 use std::{num::NonZeroU64, sync::Arc};
 
-use crate::model::{constraint::multimodal::Constraint, state::MultimodalStateMapping};
+use crate::model::constraint::multimodal::Constraint;
+use bambam_core::model::state::MultimodalStateMapping;
 
 #[derive(Debug)]
 pub struct MultimodalConstraintEngine {
