+++ rust/bambam/src/model/mod.rs
@@ -3,5 +3,4 @@ pub mod constraint;
 pub mod input_plugin;
 pub mod label;
 pub mod output_plugin;
-pub mod state;
 pub mod traversal;
